Ch. 17 §02: correct numbering and content for Listings 17-13 and 17-14
