+++ listings/ch17-async-await/listing-17-13-orig/src/main.rs
@@ -1,10 +1,15 @@
-use std::time::Duration;
+use std::{
+    future::Future,
+    pin::{pin, Pin},
+    time::Duration,
+};
 
 fn main() {
     trpl::block_on(async {
         let (tx, mut rx) = trpl::channel();
 
-        let tx_fut = async {
+        let tx1 = tx.clone();
+        let tx1_fut = pin!(async move {
             let vals = vec![
                 String::from("hi"),
                 String::from("from"),
@@ -13,19 +18,18 @@ fn main() {
             ];
 
             for val in vals {
-                tx.send(val).unwrap();
+                tx1.send(val).unwrap();
                 trpl::sleep(Duration::from_secs(1)).await;
             }
-        };
+        });
 
-        let rx_fut = async {
-            while let Some(received) = rx.recv().await {
-                println!("Got: {received}");
+        let rx_fut = pin!(async {
+            while let Some(value) = rx.recv().await {
+                println!("received '{value}'");
             }
-        };
+        });
 
-        // ANCHOR: updated
-        let tx_fut2 = async {
+        let tx_fut = pin!(async move {
             let vals = vec![
                 String::from("more"),
                 String::from("messages"),
@@ -37,9 +41,13 @@ fn main() {
                 tx.send(val).unwrap();
                 trpl::sleep(Duration::from_secs(1)).await;
             }
-        };
+        });
 
-        trpl::join3(tx_fut, tx_fut2, rx_fut).await;
-        // ANCHOR_END: updated
+        // ANCHOR: here
+        let futures: Vec<Pin<Box<dyn Future<Output = ()>>>> =
+            vec![Box::pin(tx1_fut), Box::pin(rx_fut), Box::pin(tx_fut)];
+        // ANCHOR_END: here
+
+        trpl::join_all(futures).await;
     });
 }
