+++ listings/ch17-async-await/listing-17-13/src/main.rs
@@ -1,15 +1,11 @@
-use std::{
-    future::Future,
-    pin::{pin, Pin},
-    time::Duration,
-};
+use std::time::Duration;
 
 fn main() {
     trpl::block_on(async {
         let (tx, mut rx) = trpl::channel();
 
-        let tx1 = tx.clone();
-        let tx1_fut = pin!(async move {
+        // ANCHOR: futures
+        let tx_fut = async {
             let vals = vec![
                 String::from("hi"),
                 String::from("from"),
@@ -18,36 +14,20 @@ fn main() {
             ];
 
             for val in vals {
-                tx1.send(val).unwrap();
-                trpl::sleep(Duration::from_secs(1)).await;
-            }
-        });
-
-        let rx_fut = pin!(async {
-            while let Some(value) = rx.recv().await {
-                println!("received '{value}'");
-            }
-        });
-
-        let tx_fut = pin!(async move {
-            let vals = vec![
-                String::from("more"),
-                String::from("messages"),
-                String::from("for"),
-                String::from("you"),
-            ];
-
-            for val in vals {
                 tx.send(val).unwrap();
                 trpl::sleep(Duration::from_secs(1)).await;
             }
-        });
+        };
 
-        // ANCHOR: here
-        let futures: Vec<Pin<Box<dyn Future<Output = ()>>>> =
-            vec![Box::pin(tx1_fut), Box::pin(rx_fut), Box::pin(tx_fut)];
-        // ANCHOR_END: here
+        let rx_fut = async {
+            // ANCHOR: loop
+            while let Some(value) = rx.recv().await {
+                println!("received '{value}'");
+            }
+            // ANCHOR_END: loop
+        };
 
-        trpl::join_all(futures).await;
+        trpl::join(tx_fut, rx_fut).await;
+        // ANCHOR_END: futures
     });
 }
